Made improvements to core scripts. Finalized some scripts from the sample and moved them into core. Split Appearance and Transform handlers into separate behaviors. Added folders for better organization.
