+++ Assets/MixedRealityAzure/LUIS/Scripts/Context.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: de8d4989ea4aea3438cb679b4986522e
+folderAsset: yes
+timeCreated: 1524157088
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
