+++ Assets/MixedRealityAzure/LUIS/Scripts/Intent.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: ed9e86a5d15c84d47bbc6627ea889a64
+folderAsset: yes
+timeCreated: 1524157130
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
