+++ Assets/MixedRealityAzure/LUIS/Scripts/Intent/DebugIntentHandler.cs
@@ -25,24 +25,50 @@
 
 using Microsoft.Cognitive.LUIS;
 using Microsoft.MR.LUIS;
-using System.Collections;
+using System;
 using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
 using UnityEngine;
 
-public class LUISTester : MonoBehaviour
+/// <summary>
+/// A global intent handler that prints information about all received intents.
+/// </summary>
+public class DebugIntentHandler : IIntentHandler
 {
-    public LuisManager luisManager;
-    public string testUtterence = "Set the box to blue";
+	#region Public Methods
+	/// <summary>
+	/// <inheritdoc/>
+	/// </summary>
+	public bool CanHandle(string intentName)
+	{
+		//This is the debug handler, we want it to be able to act on any intent.
+		return true;
+	}
 
-    // Use this for initialization
-    async void Start()
-    {
-        luisManager.EntityResolvers.Add(new EntityMetaDataResolver());
-        luisManager.IntentHandlers.Add(new DebugHandler());
-
-        if (!string.IsNullOrEmpty(testUtterence))
-        {
-            await luisManager.PredictAndHandle(testUtterence);
-        }
-    }
+	/// <summary>
+	/// <inheritdoc/>
+	/// </summary>
+	public void Handle(Intent intent, LuisMRResult result)
+	{
+		//Build up a string of information about the result we got from LUIS
+		StringBuilder sb = new StringBuilder();
+		sb.AppendLine("Utterance: " + result.Context.PredictionText);
+		sb.AppendLine("Intent: " + intent.Name);
+		sb.AppendLine("  Score: " + intent.Score.ToString("P"));
+		sb.AppendLine("Entities: ");
+		foreach (string entityKey in result.PredictionResult.Entities.Keys)
+		{
+			sb.AppendLine("  Entity: " + entityKey);
+			foreach (Entity e in result.PredictionResult.Entities[entityKey])
+			{
+				sb.AppendLine("    Value:  " + e.Value);
+				sb.AppendLine("    Score:  " + e.Score);
+			}
+		}
+		//Then write it to the console
+		Debug.Log(sb.ToString());
+	}
+	#endregion // Public Methods
 }
