+++ Assets/MixedRealityAzure/LUIS/Scripts/Resolve.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: d02aacc3cab885b4485f41a16041d4b1
+folderAsset: yes
+timeCreated: 1524157155
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
