+++ Assets/MixedRealityAzure/LUIS/Scripts/Route.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 081e1d33da47fc846992b500054cc340
+folderAsset: yes
+timeCreated: 1524157207
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
